u-boot-dey: restore rootfs variables when rolling back
When the update process is completed several variables related to the
rootfs to mount are updated.

Restore them all to their previous active system values when the new
system is unable to boot for 'bootlimit' attempts like we currently do
with 'active_system'.

Use 'setexpr' for 'mmcpart' variable to ensure it is stored in decimal
format.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>